Validate AST of transformed files
After running all transforms on a file, this validates that the
resulting file can be parsed by the AST module.  For files with bad
syntax or other errors, generate a useful error message on stderr, and
skip processing those hunks.  Also tracks the exceptions generated, and
builds a list of all exceptions on the Query object, so they can be
tracked and/or inspected after the query is executed.