Switch validation to use fissix instead of ast.
Because ast is from the host python, it rejects all kinds of valid (to some
version of Python) syntax that we might not even have modified in the
transform.  This uses fissix, which I believe is strict enough to catch the
typical errors like mismatched parens, extra commas, or bad indentation.

Fixes #77